// frontend/src/lib/api/models/ConnectionResponse.ts
/* generated using openapi-typescript-codegen -- do not edit */
/* istanbul ignore file */
/* tslint:disable */
/* eslint-disable */
import type { ArtifactResponseForConnections } from './ArtifactResponseForConnections';
export type ConnectionResponse = {
    source: ArtifactResponseForConnections;
    target: ArtifactResponseForConnections;
};